expand mutant coverage for link normalization
Extract markdown finalization and local link rewriting into a dedicated helper module so mutant can exercise path resolution and malformed-link cleanup directly. Add focused tests for registry resolution, relative path rewriting, constant-reference detection, and final markdown normalization, and accept the simplifications mutant proved redundant.
