+++ lib/yard-markdown.rb
@@ -5,6 +5,7 @@ require_relative "yard/markdown/anchor_component_helper"
 require_relative "yard/markdown/aref_helper"
 require_relative "yard/markdown/documentation_helper"
 require_relative "yard/markdown/heading_helper"
+require_relative "yard/markdown/link_normalization_helper"
 require_relative "yard/markdown/method_presentation_helper"
 require_relative "yard/markdown/object_listing_helper"
 require_relative "yard/markdown/section_assembly_helper"
